requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
mysql-connector-python==8.0.26
python-dotenv==0.19.0
requests==2.26.0
googlemaps==4.6.0
python-socketio==5.4.0
Flask-SocketIO==5.1.1
google-cloud-language==2.4.1
google-cloud-translate==3.7.1
